qcacld-2.0: Fix OOB read in scan_id get
Propagation from qcacld-3.0 to qcacld-2.0
In wma_extscan_find_unique_scan_ids() the TLV structures
for param_buf are pulled from the WMI message.
wma_extscan_find_unique_scan_ids parses the data (param_buf)
which is obtained from the firmware.
This parsing logic of rssi_list does not consider the
size of the list and thus results in an OOB access.

Fix is loop for the num_rssi_list and not the num of entries

Change-Id: Icf79b59a17b66ac858222b79589641787022572d
CRs-Fixed: 2350711
Signed-off-by: mortezadadgar <mortezadadgar97@gmail.com>